+++ changelog.md
@@ -1,6 +1,14 @@
 # Changelog
 
-#### Unreleased Changes
+#### Version 1.6 - 17. October 2019
+- Now writes overlay data to a json next to the cleopetra.jar. This can be disabled in RLBotSettings tab. - NicEastvillage
+- Improved the RLBot runner process. - NicEastvillage
+    - The RLBot console window (RLBot runner) now does not need to be shut down after each match. #88
+    - CleoPetra issues the RLBot runner to start and stop new matches through socket communication.
+    - RLBot.exe is not shut down between each match, which means:
+        - Skyborg's overlay will work properly.
+        - Rendering and bot percentages does not have to be toggled each match.
+- Improved how round robin decides the top teams of the stage. #92 - NicEastvillage
 - Added button for auto-naming teams based on its bots. #79 - NicEastvillage
 
 #### Version 1.5.1 - 3. October 2019